+++ src/Properties/AssemblyInfo.cs
@@ -46,4 +46,4 @@ using System.Runtime.InteropServices;
 // - Minor should be for features and critical bugs
 // - Revision is Bug fixes, ad-hoc patches, any minor changes.
 // - Build is an (automated)incremented build number.
-[assembly: AssemblyVersion("15.0.2.0")]
+[assembly: AssemblyVersion("15.0.3.0")]